adding flow control validation

+++ app/Http/Controllers/MarcaController.php
@@ -47,6 +47,9 @@ class MarcaController extends Controller
     public function show($id)
     {
         $marca = $this->marca->find($id);
+        if ($marca === null) {
+            return ['erro' => 'Recurso pesquisado não existe'];
+        }
         return $marca;
     }
 
@@ -61,6 +64,9 @@ class MarcaController extends Controller
     public function update(Request $request, $id)
     {
         $marca = $this->marca->find($id);
+        if ($marca === null) {
+            return ['erro' => 'Impossível realizar a atualização, o recurso solicitado não existe.'];
+        }
 
         $marca->update($request->all());
 
@@ -76,6 +82,9 @@ class MarcaController extends Controller
     public function destroy($id)
     {
         $marca = $this->marca->find($id);
+        if ($marca === null) {
+            return ['erro' => 'Impossível realizar a exclusão, o recurso solicitado não existe.'];
+        }
 
         $marca->delete();
